refactor: remove under score from package names
