+++ src/test/java/carracing/RacingPlayerTest.java
@@ -1,7 +1,7 @@
-package car_racing;
+package carracing;
 
-import car_racing.model.Car;
-import car_racing.model.RacingPlayer;
+import carracing.model.Car;
+import carracing.model.RacingPlayer;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
